Add .operator.latest_reporting_from_{file,platform}
- Move from .transport.report and generalize.
